requirements.txt
django==5.0.2
python-dotenv==1.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.12.1
sqlparse==0.5.3
